refactoring the imports of the AbstractFacade
