+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/services/entrepriseprofile/ProjectManagerFacade.java
@@ -9,7 +9,7 @@ import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import tn.esprit.overpowered.byusforus.entities.entrepriseprofile.ProjectManager;
-import tn.esprit.overpowered.byusforus.services.quiz.AbstractFacade;
+import tn.esprit.overpowered.byusforus.entities.util.AbstractFacade;
 
 /**
  *
